Fetch the Chrome preload list at start of scan
This commit uses the `init` hook on the TLS scanner to fetch the Chrome
preload list at the start of the scan, store it in a global variable on
the TLS scanner module, and re-uses it for each subsequent scan.

Therefore, we now only download the Chrome preload list each time `scan
--scanners=tls` is run, instead once for each domain scanned. This
avoids the overhead of unnecessary re-downloads of the Chrome preload
list (since it changes infrequently) while keeping the data fresh for
each run of the scanner.