Unpin apt dependencies
Pinning the versions of the apt dependencies in the Dockerfile causes
periodic breakage of `docker-compose build` as the packages get updated.
Since we are using a stable release of a long term service distribution
of Linux, it should be safe to rely on the distribution's maintainers to
keep packages stable and compatible with each other; therefore, pinning
the apt dependneices is unnecessary, and unpinning them prevents
periodic and annoying breakage.