Use the stable branch of ssllabs-scan
Go 1.3.3 checks out the `master` branch of GitHub repositories instead of the optionally specified default branch (a GitHub feature?). The default branch for `sllabs-scan` is `stable`, but Go uses the `master` branch which is used by `ssllabs-scan` for development. These changes switch `ssllabs-scan` to the `stable` branch and installs it.

Last week, ssllabs made a small change to the format of its API JSON output which broke `ssllabs-scan`[[1](https://github.com/ssllabs/ssllabs-scan/issues/415)]. A fix has already been committed to the `stable` branch, but not to the `master` branch, so `domain-scan` was not receiving any output from `ssllabs-scan`.