Add check for Content-Security-Policy header
As a response to the BrowseAloud hack which loaded a JS cryptocurrency mining malware, this commit adds a
new scanner to check for the existance of the Content-Security-Policy header which seeks to limit the
ability of dangerous content from execution.

This is an initial version that merely checks for the canonical header as currently implemented by common
browsers.