Adding two output fields for the sslyze scanner.
The first field, "Is Symantec Cert" is a boolean field that denotes
whether the certificate was issued by Symantec.

If "Is Symantec Cert" is True, then the second field, "Symantec
Distrust Date", will contain either the string "MARCH_2018" or the
string "SEPTEMBER_2018" depending on when the SYmantec certificate was
issued.