Making it so that the logging level can be adjusted when running in Lambda
In the case of AWS Lambda, the root logger is used BEFORE our Lambda
handler runs, and this creates a default handler that goes to the
console.  Once logging has been configured, calling
logging.basicConfig() has no effect.  We can get around this by
removing any root handlers (if present) before calling
logging.basicConfig().  This unconfigures logging and allows --debug
to affect the logging level that appears in the CloudWatch logs.

See
https://stackoverflow.com/questions/1943747/python-logging-before-you-run-logging-basicconfig
and
https://stackoverflow.com/questions/37703609/using-python-logging-with-aws-lambda
for more details.