Add a generic post-scan hook
This allows us to store data in the fast cache in a format that is
most efficient for the particular scanner being run, while also
keeping the scanner-specific logic out of the scan CLI code.

At the moment I am using this hook to update the values in the fast
cache.

Note that the key associated with the fast cache value in the
environment dict is in utils/__init__.py so that the same value can be
referenced in scan as well as the individual scanners.