Add redirect information to DAP scan.
See: https://github.com/18F/site-scanning/issues/722

Why: We want to add redirect information to the DAP scan.

How: In the scanners/dap.js file I added a few things:

- On any outbound request from the page (`page.on("request", ...)`),
look to see if it is a navigation request. For now, I am ignoring HTTPS
upgrade redirects (usually a response to an HSTS header). I
think this is okay, because domain-scan does not add the protocol to
requests, so we _always_ redirect at least once from the domain scan's
perspective. Any requests detected are added to the `redirects_to`
array.

- Detect the final URL by awaiting `networkidle2`. I'm actually not sure
which heuristic we should use (`networkidle0` is the other one). From
[this response](https://github.com/puppeteer/puppeteer/issues/1552#issuecomment-350954419)
 think that we're safer with `networkidle2`. Then grab the href to
 determine the `final_url` field. Compare that with the requested domain
 to determine `final_url_in_same_domain`.

Tags: dap, puppeteer